requirements.txt
streamlit>=1.35.0
anthropic>=0.28.0
python-dotenv>=1.0.0
PyMuPDF>=1.24.0
python-docx>=1.1.0
langdetect>=1.0.9
